export log record - make optional; restructure queries

+++ civicrm/custom/php/CRM/Contact/Form/Task/ExportDistrict.php
@@ -80,6 +80,8 @@ class CRM_Contact_Form_Task_ExportDistrict extends CRM_Contact_Form_Task {
                     array( 'id' => 'locType' )
                     );
 
+        $this->addElement('checkbox', 'includeLog', ts('Include most recent log activity date'));
+
         $this->addDefaultButtons( 'Export District' );
 
         $defaults['locType'] = 0;
@@ -98,6 +100,7 @@ class CRM_Contact_Form_Task_ExportDistrict extends CRM_Contact_Form_Task {
         $params = $this->controller->exportValues( $this->_name );
         $avanti_job_id = ( $params['avanti_job_id'] ) ? 'avanti-'.$params['avanti_job_id'].'_' : '';
         $loc_type      = $params['locType'];
+		$include_log   = $params['includeLog'];
 
         //get instance name (strip first element from url)
         $instance = substr( $_SERVER['HTTP_HOST'], 0, strpos( $_SERVER['HTTP_HOST'], '.' ) );
@@ -132,16 +135,20 @@ class CRM_Contact_Form_Task_ExportDistrict extends CRM_Contact_Form_Task {
         $dao = CRM_Core_DAO::executeQuery( $sql, CRM_Core_DAO::$_nullArray );
         
         //4874
-        $logTable = createLogTable( $rnd );
+		if ( $include_log ) {
+            $logTable = createLogTable( $rnd );
+		}
 
         $sql = "SELECT c.id, c.first_name, c.middle_name, c.last_name, c.suffix_id, ";
         $sql .= "street_number, street_number_suffix, street_name, street_unit, street_address, supplemental_address_1, supplemental_address_2, city, state_province_id, postal_code, postal_code_suffix, ";
         $sql .= "c.birth_date, c.gender_id, phone, ";
         $sql .= "town_52, LPAD(ward_53,2,'0') as ward_53, LPAD(election_district_49,3,'0') as election_district_49, LPAD(congressional_district_46,2,'0') as congressional_district_46, LPAD(ny_senate_district_47,2,'0') as ny_senate_district_47, LPAD(ny_assembly_district_48,3,'0') as ny_assembly_district_48, LPAD(school_district_54,3,'0') as school_district_54, LPAD(county_50,2,'0') as county_50, ";
-        $sql .= "email, a.location_type_id, is_deleted, a.id AS address_id, di.id AS districtinfo_id, p.id AS phone_id, e.id AS email_id, ";
+        $sql .= "email, a.location_type_id, is_deleted, a.id AS address_id, di.id AS districtinfo_id, p.id AS phone_id, e.id AS email_id ";
         
         //4874
-        $sql .= "lt.mod_date AS last_modified_date ";
+		if ( $include_log ) {
+            $sql .= ", lt.mod_date AS last_modified_date ";
+		}
 
         $sql .= " FROM civicrm_contact c ";
         $sql .= " INNER JOIN tmpExport$rnd t on t.id=c.id ";
@@ -151,7 +158,9 @@ class CRM_Contact_Form_Task_ExportDistrict extends CRM_Contact_Form_Task {
         $sql .= " LEFT JOIN civicrm_email e on e.contact_id=c.id AND e.is_primary=1 ";
     
         //4874 - include last log record timestamp
-        $sql .= " LEFT JOIN $logTable lt ON c.id = lt.cid ";
+		if ( $include_log ) {
+            $sql .= " LEFT JOIN $logTable lt ON c.id = lt.cid ";
+		}
 
         $sql .= " ORDER BY CASE WHEN c.gender_id=2 THEN 1 WHEN c.gender_id=1 THEN 2 WHEN c.gender_id=4 THEN 3 ELSE 999 END, ";
         $sql .= " IFNULL(c.birth_date, '9999-01-01');";
@@ -236,8 +245,10 @@ class CRM_Contact_Form_Task_ExportDistrict extends CRM_Contact_Form_Task {
         $dao = CRM_Core_DAO::executeQuery( $sql, CRM_Core_DAO::$_nullArray );
         
         //get rid of log table
-        $sql = "DROP TABLE $logTable;";
-        //$dao = CRM_Core_DAO::executeQuery( $sql, CRM_Core_DAO::$_nullArray );
+		if ( $include_log ) {
+            $sql = "DROP TABLE $logTable;";
+            $dao = CRM_Core_DAO::executeQuery( $sql, CRM_Core_DAO::$_nullArray );
+		}
 
         $url = "http://".$_SERVER['HTTP_HOST'].'/nyss_getfile?file='.$filename;
         $urlclean = urlencode( $url );
@@ -315,53 +326,56 @@ function getStates() {
 //create table with only the most recent log entry for each contact
 function createLogTable( $rnd ) {
 
-    $tblIDs = "tmpExport$rnd";
-    $tblLog = "tmpLog$rnd";
+    $tblIDs       = "tmpExport$rnd";
+    $tblLog       = "tmpLog$rnd";
+    $tblLogDedupe = "tmpLogDedupe$rnd";
 
-    $sql = "CREATE TABLE $tblLog ( id int not null auto_increment, primary key(id), cid int not null, mod_date date );";
+    $sql = "CREATE TABLE $tblLog ( cid int not null, mod_date date );";
     $dao = CRM_Core_DAO::executeQuery( $sql, CRM_Core_DAO::$_nullArray );
 
-    //insert most recent contact log
+    //insert contact log
     //skip source as that will only be staff
     $sql = "INSERT INTO $tblLog (cid, mod_date)
-            SELECT * FROM 
-               ( SELECT c.id as cid, cl.modified_date as mod_date
-                 FROM civicrm_log cl
-                 JOIN $tblIDs c
-                   ON cl.entity_id = c.id
-                 WHERE cl.entity_table = 'civicrm_contact'
-                 ORDER BY cl.modified_date DESC, cl.id DESC ) as recentLog
-            GROUP BY cid;";
+            SELECT c.id as cid, cl.modified_date as mod_date
+            FROM civicrm_log cl
+            JOIN $tblIDs c
+              ON cl.entity_id = c.id
+            WHERE cl.entity_table = 'civicrm_contact';";
     $dao = CRM_Core_DAO::executeQuery( $sql, CRM_Core_DAO::$_nullArray );
     
-    //insert most recent activities
+    //insert activities
     //only concerned with target contact
     $sql = "INSERT INTO $tblLog (cid, mod_date)
-            SELECT * FROM
-               ( SELECT c.id as cid, cl.modified_date as mod_date
-                 FROM civicrm_log cl
-                 JOIN civicrm_activity_target cat
-                   ON cl.entity_id = cat.activity_id
-                 JOIN $tblIDs c
-                   ON cat.target_contact_id = c.id
-                 WHERE cl.entity_table = 'civicrm_activity'
-                 ORDER BY cl.modified_date DESC, cl.id DESC ) as recentLog
-            GROUP BY cid;"; 
+            SELECT c.id as cid, cl.modified_date as mod_date
+            FROM civicrm_log cl
+            JOIN civicrm_activity_target cat
+              ON cl.entity_id = cat.activity_id
+            JOIN $tblIDs c
+              ON cat.target_contact_id = c.id
+            WHERE cl.entity_table = 'civicrm_activity';"; 
     $dao = CRM_Core_DAO::executeQuery( $sql, CRM_Core_DAO::$_nullArray );
 
-    //collapse resulting data
-    $sql = "DELETE FROM $tblLog
-            WHERE id IN ( SELECT id FROM
-                            ( SELECT id, cid
-                              FROM $tblLog
-                              ORDER BY mod_date ASC ) as oldLog
-                          GROUP BY cid
-                          HAVING count(id) > 1 );";
-    $dao = CRM_Core_DAO::executeQuery( $sql, CRM_Core_DAO::$_nullArray );
+    //collapse resulting data in a separate table
+	$sql = "CREATE TABLE $tblLogDedupe LIKE $tblLog;";
+	$dao = CRM_Core_DAO::executeQuery( $sql, CRM_Core_DAO::$_nullArray );
+	
+	$sql = "ALTER TABLE $tblLogDedupe ADD UNIQUE ( cid );";
+	$dao = CRM_Core_DAO::executeQuery( $sql, CRM_Core_DAO::$_nullArray );
+	
+	$sql = "INSERT INTO $tblLogDedupe
+	        SELECT *
+			FROM $tblLog
+			ORDER BY cid, mod_date ASC
+			ON DUPLICATE KEY UPDATE mod_date = VALUES(mod_date)
+			;";
+	$dao = CRM_Core_DAO::executeQuery( $sql, CRM_Core_DAO::$_nullArray );
+	
+	//now drop first temp table
+	CRM_Core_DAO::executeQuery("DROP TABLE $tblLog;");
 
     //CRM_Core_Error::debug('tblIDs',$tblIDs);
     //CRM_Core_Error::debug('tblLog',$tblLog);
     //exit();
 
-    return $tblLog;
+    return $tblLogDedupe;
 } //createLogTable
